backend/cv_parser/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
pydantic
email-validator
python-multipart